Add Snackbar provider.
